+++ packages/plugin-sql/src/schema/index.ts
@@ -12,18 +12,4 @@ export { participantTable } from './participant';
 export { relationshipTable } from './relationship';
 export { roomTable } from './room';
 export { worldTable } from './world';
-export const taskTable = pgTable('tasks', {
-  id: uuid('id').primaryKey().defaultRandom(),
-  name: text('name').notNull(),
-  description: text('description'),
-  roomId: uuid('room_id'),
-  worldId: uuid('world_id'),
-  entityId: uuid('entity_id'),
-  agentId: uuid('agent_id').notNull(),
-  tags: text('tags')
-    .array()
-    .default(sql`'{}'::text[]`),
-  metadata: jsonb('metadata').default(sql`'{}'::jsonb`),
-  createdAt: timestamp('created_at', { withTimezone: true }).defaultNow(),
-  updatedAt: timestamp('updated_at', { withTimezone: true }).defaultNow(),
-});
+export { taskTable } from './tasks';
