+++ packages/plugin-starter/package.json
@@ -28,8 +28,8 @@
     "dist"
   ],
   "dependencies": {
-    "@elizaos/cli": "workspace:*",
-    "@elizaos/core": "workspace:*",
+    "@elizaos/cli": "^1.0.0-beta.41",
+    "@elizaos/core": "^1.0.0-beta.41",
     "zod": "3.24.2"
   },
   "devDependencies": {
